% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcalibrater.R
\name{get_shap_scores}
\alias{get_shap_scores}
\title{get_shap_scores}
\usage{
get_shap_scores(hp_data_in, bst)
}
\arguments{
\item{hp_data_in}{the pv_survey dataset}

\item{bst}{xgboost model from get_boosted_tree_model()}
}
\value{
dataframe long format
}
\description{
get_shap_scores
}
