% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_qanda_calibrate.R
\docType{data}
\name{hp_qanda_calibrate}
\alias{hp_qanda_calibrate}
\title{hp_qanda_calibrate}
\format{
A data frame with 256 rows and 4 variables:
\describe{
\item{\code{question}}{character survey question}
\item{\code{question_code}}{character question code}
\item{\code{response}}{character reponse}
\item{\code{response_code}}{double response code}
}
}
\usage{
hp_qanda_calibrate
}
\description{
a table of response options for each hp_survey question
}
\details{
2024 Li et al
}
\keyword{datasets}
