% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_questions.R
\docType{data}
\name{hp_questions}
\alias{hp_questions}
\title{hp_questions}
\format{
A data frame with 39 rows and 2 variables:
\describe{
\item{\code{question}}{character survey question}
\item{\code{question_code}}{character question code}
}
}
\usage{
hp_questions
}
\description{
a table of response options for each hp_survey question
}
\details{
2024 Li et al
}
\keyword{datasets}
