% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hp_questions_calibrate.R
\docType{data}
\name{hp_questions_calibrate}
\alias{hp_questions_calibrate}
\title{hp_questions_calibrate}
\format{
A data frame with 40 rows and 2 variables:
\describe{
\item{\code{question}}{character survey question}
\item{\code{question_code}}{character question code}
}
}
\usage{
hp_questions_calibrate
}
\description{
a table of response options for each hp_survey question
}
\details{
2024 Li et al
}
\keyword{datasets}
