% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcalibrater.R
\name{recode_bill}
\alias{recode_bill}
\title{recode_bill}
\usage{
recode_bill(hp_data_in, n_bill)
}
\arguments{
\item{hp_data_in}{input survey data}

\item{n_bill}{number of heating bill categories}
}
\value{
a new survey dataframe with annual bills range 1~ < 900, 2~900-1350, 3~1350 to 1650, 4~1650-2100,5~> 2100
}
\description{
converts to annual bill feature (q_ab), dropping min and max bills (q14 & q15)
}
