% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microcalibrater.R
\name{transform_to_utils}
\alias{transform_to_utils}
\title{transform_to_utils}
\usage{
transform_to_utils(hp_data_in, s = 0.15, epsilon = 0.7)
}
\arguments{
\item{hp_data_in}{survey data including qsp22_7 (Likert scores for adopting pv)}

\item{s}{utility uncertainty (default 0.15)}

\item{epsilon}{degree of hypothetical bias (default 0.7, 1 = no bias). Further hypothetical bias correction is applied at ABM tuning stage.}
}
\value{
modified pv survey dataframe
}
\description{
Transforms survey Likert scores for likelihood of pv adoption (1..5) to utilities in range -1,1.
The transformation depends on s (utility uncertainty) and epsilon (an initial hypothetical bias estimate).
The error in the hypithetical bias estimate is corrected at the micro-calibration stage
}
